Add files via upload
The square of an integer refers to the result of multiplying the integer with itself once. While the cube of an integer refers to the result of multiplying the integer with itself twice. As long as you know that, you could easily solve this!



Instructions:

Input three integers and compute the cubes of each of them.
Check if the sum of the cubes are positive. If it is, print out "Positive", and if not, print out "Negative".